requirements.txt
ultralytics==8.2.95
opencv-python==4.12.0.88
numpy==1.26.4
matplotlib==3.9.2
Pillow==10.4.0
PyYAML==6.0.2
